Implement new textobject for indentation level
This implements a tetobject corresponding to the current indentation
level of the selection. It is only implemented for the "match mode"
and has no difference between inside and around (as there's no obvious
meaning) to those. It does however take a count, where count extends
the selection to "lower" levels of indentation than represented by the
current selection.

I couldn't see a sensible way to do "next" and "prev" by this
textobject. Although currently it will extend past empty lines, so
maybe it could be adjusted to consider "next" and "prev" to be the next
contiguous block of the same indentation level as the current selection,
after one or more blank lines. That may also allow for defining "around"
as the current behaviour, and use "inside" to stop at blank lines. Open
to thoughts.