Make newline detection less repetitive and more robust
It turned out we already had rope_is_line_ending as a more effective
implementation of empty line detection, as it doesn't rely on only
detecting the current document's detected line ending. Thus, it'll work
with mixed line ending files. I also adjusted the removal of the final
newline from the selection to detect the new line used on that line
specifically for the same reason.